requirements.txt
beautifulsoup4==4.10.0
python-dotenv==1.0.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
lxml==4.6.4
PySocks==1.7.1
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
